+++ evalml/automl/auto_base.py
@@ -334,14 +334,14 @@ class AutoBase:
         # if the coefficient of variance is greater than .2
         high_variance_cv = (scores.std() / scores.mean()) > .2
 
-        pipeline_class_name = trained_pipeline.__class__.__name__
         pipeline_name = trained_pipeline.name
+        pipeline_summary = trained_pipeline.summary
         pipeline_id = len(self.results['pipeline_results'])
 
         self.results['pipeline_results'][pipeline_id] = {
             "id": pipeline_id,
-            "pipeline_class_name": pipeline_class_name,
             "pipeline_name": pipeline_name,
+            "pipeline_summary": pipeline_summary,
             "parameters": dict(parameters),
             "score": score,
             "high_variance_cv": high_variance_cv,
@@ -436,7 +436,7 @@ class AutoBase:
             ascending = False
 
         rankings_df = pd.DataFrame(self.results['pipeline_results'].values())
-        rankings_df = rankings_df[["id", "pipeline_class_name", "score", "high_variance_cv", "parameters"]]
+        rankings_df = rankings_df[["id", "pipeline_name", "score", "high_variance_cv", "parameters"]]
         rankings_df.sort_values("score", ascending=ascending, inplace=True)
         rankings_df.reset_index(drop=True, inplace=True)
         return rankings_df