+++ evalml/pipelines/classification/random_forest.py
@@ -6,7 +6,7 @@ from evalml.pipelines import PipelineBase
 
 class RFClassificationPipeline(PipelineBase):
     """Random Forest Pipeline for both binary and multiclass classification"""
-    name = "Random Forest Classifier w/ One Hot Encoder + Simple Imputer + RF Classifier Select From Model"
+    _name = "Random Forest Classification Pipeline"
     model_type = ModelTypes.RANDOM_FOREST
     component_graph = ['One Hot Encoder', 'Simple Imputer', 'RF Classifier Select From Model', 'Random Forest Classifier']
     problem_types = ['binary', 'multiclass']